Output is "working"
pattern doesn't loop for output
I realized looping pattern logic needs to be applied twice, once in the LCD and once in the sequence player.  Can it be combined?  (should it be combined?)
Performance is OK with no LCD enabled -- which takes 32ms to draw.
Also, our old serial bug is back